requirements.txt
requests==2.31.0
python-dotenv==1.0.0
# wave is part of Python Standard Library, no need to include